+++ cantera18/branches/scons/Cantera/src/zeroD/FlowReactor.h
@@ -37,7 +37,7 @@ namespace CanteraZeroD {
 
 
         //-----------------------------------------------------
-        virtual int neq() { return m_nv; }
+        virtual size_t neq() { return m_nv; }
 
         virtual void initialize(doublereal t0 = 0.0);
 	virtual void evalEqs(doublereal t, doublereal* y,
